requirements.txt
numpy==1.19.3
streamlit==0.79.0
Pillow==8.1.0
https://download.pytorch.org/whl/cpu/torch-1.3.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
onnx
onnxruntime
